Support the following new open() bits....
# define O_NOATIME	0x04		/* Do not set atime.  */
# define O_DIRECTORY	0x10000		/* Must be a directory.	 */
# define O_NOFOLLOW	0x20000		/* Do not follow links.	 */

the kernel alreadys knows how to deal with the above on the
freemint-enhancements branch.

On 1.16 and earlier, O_DIRECTORY won't work as it always fails to open
directories and EISDIR is returned and fails. O_NOATIME will be ignored.

As for O_NOFOLLOW, that's purely a mintlib feature and should work fine
with 1.16 and earlier.