Prevent features.h from being "fixed" when compiling gcc
Newer gcc versions fix that header file, and place a copy in gccs
private include-fixed directory. But we need to support also older compilers,
and that prevents us from updating the file by newer versions of mintlib.

The fix just adds a space, so the pattern in fixincludes just does not match.