Update tzcode to 2025a
Note: upstream version introduced various changes in the code that i
decided to skip. One of them would be to use a type larger than int in
localtime() in order to avoid having to check for overflows. In our
case that would be long long, which is imho unacceptable since it
introduces lots of expensive operations, just to avoid corner cases
that "should not happen".

Also, that caused localtime.c to be cluttered with lots of ifdefs, and
duplicate code of it being used in strftime()

Another change was to make the argument to update_tzname_etc() and some
others const, and then introducing ugly code to cast away the constness
again.